+++ ImGui.Widgets/Image.cs
@@ -22,7 +22,7 @@ public static partial class ImGuiWidgets
 	/// <param name="textureId">The ID of the texture to display.</param>
 	/// <param name="size">The size of the image.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool Image(uint textureId, Vector2 size) => ImageImpl.Show(textureId, size, Vector4.One);
+	public static bool Image(nint textureId, Vector2 size) => ImageImpl.Show(textureId, size, Vector4.One);
 
 	/// <summary>
 	/// Displays an image with the specified texture ID, size, and color.
@@ -31,7 +31,7 @@ public static partial class ImGuiWidgets
 	/// <param name="size">The size of the image.</param>
 	/// <param name="color">The color to apply to the image.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool Image(uint textureId, Vector2 size, Vector4 color) => ImageImpl.Show(textureId, size, color);
+	public static bool Image(nint textureId, Vector2 size, Vector4 color) => ImageImpl.Show(textureId, size, color);
 
 	/// <summary>
 	/// Displays a centered image with the specified texture ID and size.
@@ -39,7 +39,7 @@ public static partial class ImGuiWidgets
 	/// <param name="textureId">The ID of the texture to display.</param>
 	/// <param name="size">The size of the image.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool ImageCentered(uint textureId, Vector2 size) => ImageImpl.Centered(textureId, size, Vector4.One);
+	public static bool ImageCentered(nint textureId, Vector2 size) => ImageImpl.Centered(textureId, size, Vector4.One);
 
 	/// <summary>
 	/// Displays a centered image with the specified texture ID, size, and color.
@@ -48,7 +48,7 @@ public static partial class ImGuiWidgets
 	/// <param name="size">The size of the image.</param>
 	/// <param name="color">The color to apply to the image.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool ImageCentered(uint textureId, Vector2 size, Vector4 color) => ImageImpl.Centered(textureId, size, color);
+	public static bool ImageCentered(nint textureId, Vector2 size, Vector4 color) => ImageImpl.Centered(textureId, size, color);
 
 	/// <summary>
 	/// Displays a centered image within a container with the specified texture ID, size, and container size.
@@ -57,7 +57,7 @@ public static partial class ImGuiWidgets
 	/// <param name="size">The size of the image.</param>
 	/// <param name="containerSize">The size of the container.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool ImageCenteredWithin(uint textureId, Vector2 size, Vector2 containerSize) => ImageImpl.CenteredWithin(textureId, size, containerSize, Vector4.One);
+	public static bool ImageCenteredWithin(nint textureId, Vector2 size, Vector2 containerSize) => ImageImpl.CenteredWithin(textureId, size, containerSize, Vector4.One);
 
 	/// <summary>
 	/// Displays a centered image within a container with the specified texture ID, size, container size, and color.
@@ -67,7 +67,7 @@ public static partial class ImGuiWidgets
 	/// <param name="containerSize">The size of the container.</param>
 	/// <param name="color">The color to apply to the image.</param>
 	/// <returns>True if the image is clicked; otherwise, false.</returns>
-	public static bool ImageCenteredWithin(uint textureId, Vector2 size, Vector2 containerSize, Vector4 color) => ImageImpl.CenteredWithin(textureId, size, containerSize, color);
+	public static bool ImageCenteredWithin(nint textureId, Vector2 size, Vector2 containerSize, Vector4 color) => ImageImpl.CenteredWithin(textureId, size, containerSize, color);
 
 	internal static class ImageImpl
 	{
@@ -77,7 +77,7 @@ public static partial class ImGuiWidgets
 		/// <param name="textureId">The ID of the texture to display.</param>
 		/// <param name="size">The size of the image.</param>
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
-		internal static bool Show(uint textureId, Vector2 size) => Show(textureId, size, Vector4.One);
+		internal static bool Show(nint textureId, Vector2 size) => Show(textureId, size, Vector4.One);
 
 		/// <summary>
 		/// Displays an image with the specified texture ID, size, and color.
@@ -88,7 +88,7 @@ public static partial class ImGuiWidgets
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
 		[SuppressMessage("Major Code Smell", "S3427:Method overloads with default parameter values should not overlap", Justification = "The no-arg overload intentionally uses Vector4.One as default, distinct from the default(Vector4) sentinel; both are needed for correct behavior.")]
 		[SuppressMessage("Major Code Smell", "S6640:Make sure that using \"unsafe\" is safe here", Justification = "Required for native ImGui/OpenGL interop; pointers are scoped to the call and not retained.")]
-		internal static bool Show(uint textureId, Vector2 size, Vector4 color = default)
+		internal static bool Show(nint textureId, Vector2 size, Vector4 color = default)
 		{
 			unsafe
 			{
@@ -111,7 +111,7 @@ public static partial class ImGuiWidgets
 		/// <param name="textureId">The ID of the texture to display.</param>
 		/// <param name="size">The size of the image.</param>
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
-		internal static bool Centered(uint textureId, Vector2 size) => Centered(textureId, size, Vector4.One);
+		internal static bool Centered(nint textureId, Vector2 size) => Centered(textureId, size, Vector4.One);
 
 		/// <summary>
 		/// Displays a centered image with the specified texture ID, size, and color.
@@ -120,7 +120,7 @@ public static partial class ImGuiWidgets
 		/// <param name="size">The size of the image.</param>
 		/// <param name="color">The color to apply to the image.</param>
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
-		internal static bool Centered(uint textureId, Vector2 size, Vector4 color)
+		internal static bool Centered(nint textureId, Vector2 size, Vector4 color)
 		{
 			bool clicked = false;
 			using (new Alignment.Center(size))
@@ -138,7 +138,7 @@ public static partial class ImGuiWidgets
 		/// <param name="size">The size of the image.</param>
 		/// <param name="containerSize">The size of the container.</param>
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
-		internal static bool CenteredWithin(uint textureId, Vector2 size, Vector2 containerSize) => CenteredWithin(textureId, size, containerSize, Vector4.One);
+		internal static bool CenteredWithin(nint textureId, Vector2 size, Vector2 containerSize) => CenteredWithin(textureId, size, containerSize, Vector4.One);
 
 		/// <summary>
 		/// Displays a centered image within a container with the specified texture ID, size, container size, and color.
@@ -148,7 +148,7 @@ public static partial class ImGuiWidgets
 		/// <param name="containerSize">The size of the container.</param>
 		/// <param name="color">The color to apply to the image.</param>
 		/// <returns>True if the image is clicked; otherwise, false.</returns>
-		internal static bool CenteredWithin(uint textureId, Vector2 imageSize, Vector2 containerSize, Vector4 color)
+		internal static bool CenteredWithin(nint textureId, Vector2 imageSize, Vector2 containerSize, Vector4 color)
 		{
 			bool clicked = false;
 			using (new Alignment.CenterWithin(imageSize, containerSize))
